requirements.txt
Django==1.6.3
PyJWT==0.2.0
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-cors-headers==0.12
django-toolbelt==0.0.1
djangorestframework==2.3.13
gunicorn==18.0
newrelic==2.18.1.15
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
